+++ backend/src/teams/teams.module.ts
@@ -2,10 +2,11 @@ import { Module } from '@nestjs/common';
 import { TeamsService } from './teams.service';
 import { TeamsController } from './teams.controller';
 import { AuditModule } from '../audit/audit.module';
+import { IsWardDirectoryConstraint } from '../common/validators/is-ward-directory.validator';
 
 @Module({
   imports: [AuditModule],
-  providers: [TeamsService],
+  providers: [TeamsService, IsWardDirectoryConstraint], // v0.33: register class-validator constraint
   controllers: [TeamsController],
   exports: [TeamsService],
 })